Fixed numbering error in RayStepper
RayStepper didn't have the numbering for locations that were out of
bounds correct, so it couldn't do the bounds detection. I added some
test cases to cover these situations. Additionally, I reduced the number
of bits needed for the propagator because we now know that the root node
shouldn't ever be empty.